[REFACTOR/#12] 폴더명 변경
