+++ app/src/main/AndroidManifest.xml
@@ -61,12 +61,12 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".presentation.storage.CreateNewCaveActivity"
+            android:name=".presentation.cavecreate.CreateNewCaveActivity"
             android:windowSoftInputMode="adjustResize"
-            android:exported="true">
+            android:exported="false">
         </activity>
         <activity
-            android:name=".presentation.storage.SeeNewCaveActivity"
+            android:name=".presentation.cavecreate.SeeNewCaveActivity"
             android:windowSoftInputMode="adjustResize"
             android:exported="false"/>
     </application>
