+++ app/src/main/java/com/growthook/aos/presentation/cavecreate/CreateNewCaveViewModel.kt
@@ -1,4 +1,4 @@
-package com.growthook.aos.presentation.storage
+package com.growthook.aos.presentation.cavecreate
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MediatorLiveData
