+++ app/src/main/java/com/growthook/aos/presentation/cavecreate/SeeNewCaveActivity.kt
@@ -1,10 +1,10 @@
-package com.growthook.aos.presentation.storage
+package com.growthook.aos.presentation.cavecreate
 
 import android.os.Bundle
 import androidx.activity.viewModels
-import com.growthook.aos.data.entity.NewStorage
+import com.growthook.aos.presentation.cavecreate.model.NewCaveIntent
 import com.growthook.aos.databinding.ActivitySeeNewCaveBinding
-import com.growthook.aos.presentation.storage.CreateNewCaveActivity.Companion.NEW_STORAGE
+import com.growthook.aos.presentation.cavecreate.CreateNewCaveActivity.Companion.NEW_STORAGE
 import com.growthook.aos.util.base.BaseActivity
 import com.growthook.aos.util.extension.getParcelable
 import dagger.hilt.android.AndroidEntryPoint
@@ -26,10 +26,10 @@ class SeeNewCaveActivity: BaseActivity<ActivitySeeNewCaveBinding>({
 
 
     private fun getStorageData() {
-        val newStorageData = intent.getParcelable(NEW_STORAGE, NewStorage::class.java)
+        val newCaveIntentData = intent.getParcelable(NEW_STORAGE, NewCaveIntent::class.java)
         with (binding) {
-            tvSeeStorageTitle.text = newStorageData?.name ?: ""
-            tvSeeStorageSubTitle.text = newStorageData?.introduction ?: ""
+            tvSeeStorageTitle.text = newCaveIntentData?.name ?: ""
+            tvSeeStorageSubTitle.text = newCaveIntentData?.introduction ?: ""
         }
     }
 }
